Rename chatbot page to BusinessAssistant
